+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ###### Enhancements
 
+- Add Podfile `podspec` which allows to use the dependencies of a podspec file. [#162](https://github.com/CocoaPods/CocoaPods/issues/162)
 - Check if any of the build settings defined in the xcconfig files is overridden. [#92](https://github.com/CocoaPods/CocoaPods/issues/92)
 - The Linter now checks that there are no compiler flags that disable warnings.